Implementing in and out functions
Add two input and two output functions to read and write one or two
bytes respectively (insb, insw, outb, outw). This functions are defined
in assembly and exported as global to allow its use from the C kernel
code.

Port-mapped I/O often uses a special class of CPU instructions designed
specifically for performing I/O, such as the in and out instructions
found on microprocessors based on the x86 and x86-64 architectures.
Different forms of these two instructions can copy one, two or four
bytes (outb, outw and outl, respectively) between the EAX register or
one of that register's subdivisions on the CPU and a specified I/O port
which is assigned to an I/O device. I/O devices have a separate address
space from general memory, either accomplished by an extra "I/O" pin on
the CPU's physical interface, or an entire bus dedicated to I/O. Because
the address space for I/O is isolated from that for main memory, this is
sometimes referred to as isolated I/O.

https://en.wikipedia.org/wiki/Memory-mapped_I/O