Programmable Interrupt Controller Implementation
- Added remap master PIC logic to kernel (ASM).
- Added wrapper functions in IDT (ASM) for interrupt handlers (C), this
wrappers disable interrupts, before calling the handler, and reenable
after the call, and ends with the iret instruction, neccessary after an
interruption handling.
- Set keyboard interrupt handler as example.

Interrupts are signals from a device, such as a keyboard or a hard drive
, to the CPU, telling it to immediately stop whatever it is currently
doing and do something else. For example, a keyboard controller can send
an interrupt when a character key was pressed. Then the OS can display
the character on screen immediately, even if the CPU was doing
something completely unrelated before, and return to what it was doing
afterwards. When a specific interrupt arises, the CPU looks up an entry
for that specific interrupt from a table provided by the OS. In x86
protected mode the table is called the Interrupt Descriptor Table (IDT)
and can have up to 256 entries, but the name of this table and the
maximum number of entries it can have can differ based on the CPU
architecture. After the CPU finds the entry for the interrupt, it jumps
to the code the entry points to. This code that is run in response to
the interrupt is known as a interrupt service routine (ISR) or an
interrupt handler.

The 8259 Programmable Interrupt Controller (PIC) is one of the most
important chips making up the x86 architecture. Without it, the x86
architecture would not be an interrupt driven architecture. The function
of the 8259A is to manage hardware interrupts and send them to the
appropriate system interrupt. This allows the system to respond to
devices needs without loss of time (from polling the device, for
instance).

Basically, when a key is pressed, the keyboard controller tells a device
called the Programmable Interrupt Controller, or PIC, to cause an
interrupt. Because of the wiring of keyboard and PIC, IRQ #1 is the
keyboard interrupt, so when a key is pressed, IRQ 1 is sent to the PIC.
The role of the PIC will be to decide whether the CPU should be
immediately notified of that IRQ or not and to translate the IRQ number
into an interrupt vector (i.e. a number between 0 and 255) for the CPU's
table.

References:
https://wiki.osdev.org/Interrupts
https://wiki.osdev.org/PIC