Enable paging system
Paging is a system which allows each process to see a full virtual
address space, without actually requiring the full amount of physical
memory to be available or present. 32-bit x86 processors support 32-bit
virtual addresses and 4-GiB virtual address spaces. Currently,
implementations of x86-64 have a limit of between 4 GiB and 256 TiB of
physical address space (and an architectural limit of 4 PiB of physical
address space).

The paging_new_4gb (memory/paging/paging.c) function creates the
necessary structure to handle paging, that consist of an array of 1024
32-bit unsigned integers, the page directory, each of those integers is
a combination of an page table address (bits 31 to 12) and flags
(bits 11 to 0). A page table is also an array of 1024 32-bit unsigned
integers, each of those is also a combination of a physical address and
flags. (https://wiki.osdev.org/Paging)

             %      Address      %   Flags   %
             |     Bits 31-12    | Bits 11-0 |

Please note that it is very important that this address be 4-KiB aligned.
This is needed, due to the fact that the last 12 bits of the 32-bit
value are overwritten by access bits and such.

Both tables contain 1024 4-byte entries, making them 4 KiB each. In the
page directory, each entry points to a page table. In the page table,
each entry points to a 4 KiB physical page frame. Additionally, each
entry has bits controlling access protection and caching features of the
structure to which it points. The entire system consisting of a page
directory and page tables represents a linear 4-GiB virtual memory map.

Translation of a virtual address into a physical address first involves
dividing the virtual address into three parts: the most significant 10
bits (bits 31-22, 1024 max int) specify the index of the page directory
entry, the next 10 bits (bits 21-12, 1024 max int) specify the index of
the page table entry, and the least significant 12 bits (bits 0-11,
4096 max int) specify the page offset.

In a nutshell, the resulting structure is a page directory with 1024
entries (32-bit) that points to one page table each; every page table
also has 1024 entries that points to a 4 KiB block of physical memory.

            1024 * 1024 * 4KiB = 4GB virtual address space

References:

https://stackoverflow.com/questions/29945171/difference-between-page-table-and-page-directory/29948417#29948417